Fix namespace imports for NpmPack and NpmRunScript aliases
