+++ src/Cake.Npm/NpmPackAliases.cs
@@ -1,8 +1,7 @@
 using System;
-using System.Collections.Generic;
 using Cake.Core;
 using Cake.Core.Annotations;
-using Cake.Npm.RunScript;
+using Cake.Npm.Pack;
 
 namespace Cake.Npm
 {
@@ -14,105 +13,78 @@ namespace Cake.Npm
     public static class NpmPackAliases
     {
         /// <summary>
-        /// Runs a npm script defined in the package.json from the current folder.
+        /// Creates a npm package from the current folder.
         /// </summary>
         /// <param name="context">The context.</param>
-        /// <param name="scriptName">Name of the script to execute as defined in package.json.</param>
         /// <example>
         /// <code>
         /// <![CDATA[
-        ///     NpmRunScript("hello");
+        ///     NpmPack();
         /// ]]>
         /// </code>
         /// </example>
         [CakeMethodAlias]
-        [CakeAliasCategory("Run-Script")]
-        public static void NpmRunScript(this ICakeContext context, string scriptName)
+        [CakeAliasCategory("Pack")]
+        public static void NpmPack(this ICakeContext context)
         {
             if (context == null)
             {
                 throw new ArgumentNullException(nameof(context));
             }
 
-            if (String.IsNullOrWhiteSpace(scriptName))
-            {
-                throw new ArgumentNullException(nameof(scriptName));
-            }
-            
-            context.NpmRunScript(
-                new NpmRunScriptSettings
-                {
-                    ScriptName = scriptName
-                });
+            context.NpmPack(new NpmPackSettings());
         }
 
         /// <summary>
-        /// Runs a npm script defined in the package.json from the current folder with specific arguments.
+        /// Creates a npm package from a specific source.
         /// </summary>
         /// <param name="context">The context.</param>
-        /// <param name="scriptName">Name of the script to execute as defined in package.json.</param>
-        /// <param name="arguments">Arguments to pass to the target script</param>
+        /// <param name="source">Source to pack. Can be anything that is installable by npm, like
+        /// a package folder, tarball, tarball url, name@tag, name@version, name, or scoped name.</param>
         /// <example>
         /// <code>
         /// <![CDATA[
-        ///     NpmRunScript("hello", new List<string> { "--foo=bar" });
+        ///     NpmPack("c:\mypackagesource");
         /// ]]>
         /// </code>
         /// </example>
         [CakeMethodAlias]
-        [CakeAliasCategory("Run-Script")]
-        public static void NpmRunScript(this ICakeContext context, string scriptName, IEnumerable<string> arguments)
+        [CakeAliasCategory("Pack")]
+        public static void NpmPack(this ICakeContext context, string source)
         {
             if (context == null)
             {
                 throw new ArgumentNullException(nameof(context));
             }
 
-            if (String.IsNullOrWhiteSpace(scriptName))
+            if (String.IsNullOrWhiteSpace(source))
             {
-                throw new ArgumentNullException(nameof(scriptName));
+                throw new ArgumentNullException(nameof(source));
             }
 
-            if (arguments == null)
-            {
-                throw new ArgumentNullException(nameof(arguments));
-            }
-
-            var settings = new NpmRunScriptSettings
-            {
-                ScriptName = scriptName
-            };
-
-            foreach (var argument in arguments)
-            {
-                settings.Arguments.Add(argument);
-            }
-
-            context.NpmRunScript(settings);
+            context.NpmPack(new NpmPackSettings { Source = source });
         }
 
         /// <summary>
-        /// Runs a npm script with the specified settings.
+        /// Creates a npm package using the specified settings.
         /// </summary>
         /// <param name="context">The context.</param>
         /// <param name="settings">The settings.</param>
         /// <example>
-        /// <para>Use specific log level</para>
         /// <code>
         /// <![CDATA[
         ///     var settings = 
-        ///         new NpmRunScriptSettings 
+        ///         new NpmPackSettings 
         ///         {
-        ///             ScriptName = "hello"
         ///             LogLevel = NpmLogLevel.Verbose
         ///         };
-        ///     NpmRunScript(settings);
+        ///     NpmPack(settings);
         /// ]]>
         /// </code>
         /// </example>
         [CakeMethodAlias]
-        [CakeAliasCategory("Run-Script")]
-        public static void NpmRunScript(this ICakeContext context, NpmRunScriptSettings settings)
+        [CakeAliasCategory("Pack")]
+        public static void NpmPack(this ICakeContext context, NpmPackSettings settings)
         {
             if (context == null)
             {
@@ -126,8 +98,8 @@ namespace Cake.Npm
 
             AddinInformation.LogVersionInformation(context.Log);
 
-            var packer = new NpmScriptRunner(context.FileSystem, context.Environment, context.ProcessRunner, context.Tools, context.Log);
-            packer.RunScript(settings);
+            var packer = new NpmPacker(context.FileSystem, context.Environment, context.ProcessRunner, context.Tools, context.Log);
+            packer.Pack(settings);
         }
     }
 }
